Quizzes can now be fetched from a URL
The quiz menu JSON format completely conforms to the class standard. The
quiz JSON format, however does not. Choices are placed in an array
rather than referenced individually. In addition, each quiz has an
description field that stores a short description of the quiz that is
displayed in the subtitle of each UITableViewCell.

To do:
- Saving state
- FIX: Janky content holder view