App now preserves and restores state
If the user backgrounds the app while he is in the middle of a quiz and
the system kills QuizApp, when the user opens the app again, the user's
progress on that quiz will be preserved. QuizApp also preserves the
scroll offset of the UIScrollView that presents the choices.

Updated README.